move download button in enrichment form, replce data modal by a link toward new datagird
